# Use raw bytes instead of nested map in metav1.Fields