--- a/staging/src/k8s.io/apimachinery/pkg/api/apitesting/roundtrip/compatibility.go
+++ b/staging/src/k8s.io/apimachinery/pkg/api/apitesting/roundtrip/compatibility.go
@@ -232,11 +232,7 @@ func CompatibilityTestFuzzer(scheme *runtime.Scheme, fuzzFuncs []interface{}) *f
 			field := metav1.ManagedFieldsEntry{}
 			c.Fuzz(&field)
 			if field.Fields != nil {
-				for k1 := range field.Fields.Map {
-					for k2 := range field.Fields.Map[k1].Map {
-						field.Fields.Map[k1].Map[k2] = metav1.Fields{}
-					}
-				}
+				field.Fields.Raw = []byte("{}")
 			}
 			*f = []metav1.ManagedFieldsEntry{field}
 		},