--- a/staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/fields_test.go
+++ b/staging/src/k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal/fields_test.go
@@ -31,15 +31,9 @@ import (
 func TestFieldsRoundTrip(t *testing.T) {
 	tests := []metav1.Fields{
 		{
-			Map: map[string]metav1.Fields{
-				"f:metadata": {
-					Map: map[string]metav1.Fields{
-						".":      newFields(),
-						"f:name": newFields(),
-					},
-				},
-			},
+			Raw: []byte(`{"f:metadata":{"f:name":{},".":{}}}`),
 		},
+		EmptyFields,
 	}
 
 	for _, test := range tests {
@@ -65,16 +59,9 @@ func TestFieldsToSetError(t *testing.T) {
 	}{
 		{
 			fields: metav1.Fields{
-				Map: map[string]metav1.Fields{
-					"k:{invalid json}": {
-						Map: map[string]metav1.Fields{
-							".":      newFields(),
-							"f:name": newFields(),
-						},
-					},
-				},
+				Raw: []byte(`{"k:{invalid json}":{"f:name":{},".":{}}}`),
 			},
-			errString: "invalid character",
+			errString: "ReadObjectCB",
 		},
 	}
 
@@ -97,7 +84,7 @@ func TestSetToFieldsError(t *testing.T) {
 	}{
 		{
 			set:       *fieldpath.NewSet(invalidPath),
-			errString: "Invalid type of path element",
+			errString: "invalid PathElement",
 		},
 	}