… go.sum
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1.MutatingWebhookConfiguration.json
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1.MutatingWebhookConfiguration.pb
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1.MutatingWebhookConfiguration.yaml
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1.ValidatingWebhookConfiguration.json
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1.ValidatingWebhookConfiguration.pb
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1.ValidatingWebhookConfiguration.yaml
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1beta1.MutatingWebhookConfiguration.json
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1beta1.MutatingWebhookConfiguration.pb
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1beta1.MutatingWebhookConfiguration.yaml
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1beta1.ValidatingWebhookConfiguration.json
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1beta1.ValidatingWebhookConfiguration.pb
… staging/src/k8s.io/api/testdata/HEAD/admissionregistration.k8s.io.v1beta1.ValidatingWebhookConfiguration.yaml
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.ControllerRevision.json
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.ControllerRevision.pb
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.ControllerRevision.yaml
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.DaemonSet.json
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.DaemonSet.pb
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.DaemonSet.yaml
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.Deployment.json
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.Deployment.pb
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.Deployment.yaml
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.ReplicaSet.json
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.ReplicaSet.pb
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.ReplicaSet.yaml
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.StatefulSet.json
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.StatefulSet.pb
… staging/src/k8s.io/api/testdata/HEAD/apps.v1.StatefulSet.yaml
… staging/src/k8s.io/api/testdata/HEAD/apps.v1beta1.ControllerRevision.json
… staging/src/k8s.io/api/testdata/HEAD/apps.v1beta1.ControllerRevision.pb