**What type of PR is this?**
/kind feature
/wg apply
/sig api-machinery
/priority important-soon
/cc @lavalamp 

**What this PR does / why we need it**:
Addresses the scalability concerns in https://github.com/kubernetes/enhancements/pull/1110 as a follow-up to #80497

This uses the functions defined in https://github.com/kubernetes-sigs/structured-merge-diff/pull/90 to serialize and deserialize field sets into raw extensions, and removes metav1.Fields

Will update with benchmarks

**Does this PR introduce a user-facing change?**:
```release-note
TBD
```

**Additional documentation e.g., KEPs (Kubernetes Enhancement Proposals), usage docs, etc.**:
https://github.com/kubernetes/enhancements/pull/1110