@jennybuckley 2019-07-30
This also causes us to run into https://github.com/kubernetes/kubernetes/issues/55890 on every type