@lavalamp 2019-07-30
Ouch, to fix that I think we have to duplicate RawExtension and provide a
custom schema :/

It's possible, we do it for some types (e.g, intOrString).

Alternatively, maybe we could leave the existing type for the purpose of
generating documentation, but implement all of the custom marshalling
functions so that it's effectively raw extension under the hood? Is this
feasible?

On Tue, Jul 30, 2019 at 12:37 PM Jennifer Buckley <notifications@github.com>
wrote:

> This also causes us to run into #55890
> <https://github.com/kubernetes/kubernetes/issues/55890> on every type
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/80730?email_source=notifications&email_token=AAE6BFWVATZKTQ37MKX76M3QCCJ6JA5CNFSM4IHW7YBKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD3FBXJY#issuecomment-516561831>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFSGBW7VKT66WXDZO63QCCJ6JANCNFSM4IHW7YBA>
> .
>