@lavalamp 2019-07-30
I agree that making a custom version of RawExtension is probably the way to
go; I didn't do it in my earlier PR because I was just trying to measure
things.

On Tue, Jul 30, 2019 at 1:05 PM Jennifer Buckley <notifications@github.com>
wrote:

> @lavalamp <https://github.com/lavalamp> we could also just fix #55890
> <https://github.com/kubernetes/kubernetes/issues/55890>
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/80730?email_source=notifications&email_token=AAE6BFVGUPQK75F2CGCQW5DQCCNJJA5CNFSM4IHW7YBKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD3FEX2Q#issuecomment-516574186>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFTJWBOEVZHDNKHP5XTQCCNJJANCNFSM4IHW7YBA>
> .
>