@jennybuckley 2019-07-30
@lavalamp right now I'm trying to see if I can change the existing metav1.Fields type into something like that