@lavalamp 2019-07-30
Sounds great, thanks!

On Tue, Jul 30, 2019, 1:22 PM Jennifer Buckley <notifications@github.com>
wrote:

> @lavalamp <https://github.com/lavalamp> right now I'm trying to see if I
> can change the existing metav1.Fields type into something like that
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/80730?email_source=notifications&email_token=AAE6BFSRASZAJEX7AOPV5ELQCCPJRA5CNFSM4IHW7YBKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD3FGF2A#issuecomment-516580072>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFV3ZFWGBEASPPF7FL3QCCPJRANCNFSM4IHW7YBA>
> .
>