@jennybuckley 2019-07-30
Looks like these four e2e ones are failing on every PR right now

The compatibility failure is also expected, since we are deliberately treating old data in proto different from old data in json