@jennybuckley 2019-07-30
Okay, it turns out that updating compatibility fixture data doesn't fix it completely, since we still need to account for why the decoded objects differ when coming from json and proto in an old version.

This is an alpha field and I think we can live with that difference. I see three ways around this:
- we could get rid of the difference (and make the test pass) by changing the field name from `managedFields[].fields` to something else like `managedFields[].fieldSet`
- we could get rid of the difference by making a custom proto marshaller that handles the old data too
- we could allow the difference in the test