@lavalamp 2019-07-31
How about a proto marshaller that recognizes and drops the old data? is
that difficult?

On Tue, Jul 30, 2019 at 4:23 PM Jennifer Buckley <notifications@github.com>
wrote:

> Okay, it turns out that updating compatibility fixture data doesn't fix it
> completely, since we still need to account for why the decoded objects
> differ when coming from json and proto in an old version.
>
> This is an alpha field and I think we can live with that difference. I see
> three ways around this:
>
>    - we could get rid of the difference (and make the test pass) by
>    changing the field name from managedFields[].fields to something else
>    like managedFields[].fieldSet
>    - we could get rid of the difference by making a custom proto
>    marshaller that handles the old data too
>    - we could allow the difference in the test
>
> —
> You are receiving this because you were mentioned.
> Reply to this email directly, view it on GitHub
> <https://github.com/kubernetes/kubernetes/pull/80730?email_source=notifications&email_token=AAE6BFQSLUW3IZY4RXOMNJ3QCDEPNA5CNFSM4IHW7YBKYY3PNVWWK3TUL52HS4DFVREXG43VMVBW63LNMVXHJKTDN5WW2ZLOORPWSZGOD3FS4ZA#issuecomment-516632164>,
> or mute the thread
> <https://github.com/notifications/unsubscribe-auth/AAE6BFWHZAMBEB6XSYAFQBDQCDEPNANCNFSM4IHW7YBA>
> .
>