@jennybuckley 2019-08-01
Old benchmarks:
```
$ go test k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager -bench=.
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager
BenchmarkApplyNewObject-12     	    1000	   1670551 ns/op	  500674 B/op	    7167 allocs/op
BenchmarkUpdateNewObject-12    	    1000	   1867805 ns/op	  560253 B/op	    8396 allocs/op
BenchmarkRepeatedUpdate-12     	    2000	    839920 ns/op	  221142 B/op	    2887 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager	5.793s
$ go test k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal -bench=.
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal
BenchmarkSetToFields-12    	  100000	     18517 ns/op	    8369 B/op	     124 allocs/op
BenchmarkFieldsToSet-12    	  100000	     12771 ns/op	    4688 B/op	      93 allocs/op
BenchmarkYAMLToTyped-12    	   30000	     57630 ns/op	   27408 B/op	     152 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal	6.389s
```
New benchmarks:
```
$ go test k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager -bench=.
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager
BenchmarkApplyNewObject-12     	    1000	   1650093 ns/op	  481842 B/op	    6820 allocs/op
BenchmarkUpdateNewObject-12    	    1000	   1800949 ns/op	  534794 B/op	    7901 allocs/op
BenchmarkRepeatedUpdate-12     	    2000	    659418 ns/op	  190649 B/op	    2264 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager	5.276s
$ go test k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal -bench=.
goos: linux
goarch: amd64
pkg: k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal
BenchmarkSetToFields-12    	  300000	      4322 ns/op	    1016 B/op	      14 allocs/op
BenchmarkFieldsToSet-12    	  200000	      9425 ns/op	    6417 B/op	      82 allocs/op
BenchmarkYAMLToTyped-12    	   30000	     58456 ns/op	   27416 B/op	     152 allocs/op
PASS
ok  	k8s.io/apiserver/pkg/endpoints/handlers/fieldmanager/internal	6.292s
```