@lavalamp 2019-08-01
Strange, I thought my PoC showed more improvement for the first two benchmarks, and this should be the same?