@lavalamp 2019-08-01
/lgtm
/approve

I think a followup with a bit more in the way of golden data for catching future format breakage might be good.