@lavalamp 2019-08-02
Oh I forgot for a second that this is an API change.

Hey @liggitt, @smarterclayton, fyi, we're making the API a little bit weirder as discussed at the SIG API Machinery meeting a while back.