@jennybuckley 2019-08-06
Ran some scalability tests on this, these were the results. Going to compare these more and see if the improvement is enough

```
old:

testing/density
APIResponsiveness: WARNING Top latency metric: {Resource:pods Subresource: Verb:POST Scope:namespace Latency:perc50: 383.515ms, perc90: 752.228ms, perc99: 1.242136s Count:3501}; threshold: 1s
APIResponsiveness: WARNING Top latency metric: {Resource:configmaps Subresource: Verb:POST Scope:namespace Latency:perc50: 4.872ms, perc90: 260.909ms, perc99: 1.076742s Count:24}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:binding Verb:POST Scope:namespace Latency:perc50: 340.473ms, perc90: 736.154ms, perc99: 968.789ms Count:3500}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource: Verb:DELETE Scope:namespace Latency:perc50: 247.69ms, perc90: 580.675ms, perc99: 966.046ms Count:7000}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:status Verb:PATCH Scope:namespace Latency:perc50: 106.528ms, perc90: 393.084ms, perc99: 878.509ms Count:17635}; threshold: 1s

testing/load
APIResponsiveness: Top latency metric: {Resource:nodes Subresource: Verb:LIST Scope:cluster Latency:perc50: 28.07ms, perc90: 93.536ms, perc99: 1.099665s Count:178}; threshold: 30s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:binding Verb:POST Scope:namespace Latency:perc50: 11.568ms, perc90: 383.024ms, perc99: 960.288ms Count:3251}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource: Verb:POST Scope:namespace Latency:perc50: 14.699ms, perc90: 454.426ms, perc99: 944.308ms Count:3251}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:replicasets Subresource: Verb:DELETE Scope:namespace Latency:perc50: 2.904ms, perc90: 30.796ms, perc99: 758.386ms Count:328}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:status Verb:PATCH Scope:namespace Latency:perc50: 13.661ms, perc90: 141.517ms, perc99: 639.23ms Count:18043}; threshold: 1s

new:

testing/density
APIResponsiveness: Top latency metric: {Resource:nodes Subresource: Verb:LIST Scope:cluster Latency:perc50: 14.22ms, perc90: 108.716ms, perc99: 1.013075s Count:97}; threshold: 30s
APIResponsiveness: Top latency metric: {Resource:pods Subresource: Verb:POST Scope:namespace Latency:perc50: 263.879ms, perc90: 556.115ms, perc99: 879.549ms Count:3500}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:binding Verb:POST Scope:namespace Latency:perc50: 256.942ms, perc90: 513.08ms, perc99: 742.577ms Count:3500}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:status Verb:PATCH Scope:namespace Latency:perc50: 49.876ms, perc90: 292.037ms, perc99: 583.91ms Count:18518}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource: Verb:GET Scope:namespace Latency:perc50: 35.935ms, perc90: 226.091ms, perc99: 492.907ms Count:25344}; threshold: 1s

testing/load
APIResponsiveness: Top latency metric: {Resource:nodes Subresource: Verb:LIST Scope:cluster Latency:perc50: 13.842ms, perc90: 56.789ms, perc99: 345.596ms Count:179}; threshold: 30s
APIResponsiveness: Top latency metric: {Resource:pods Subresource: Verb:DELETE Scope:namespace Latency:perc50: 11.827ms, perc90: 116.55ms, perc99: 339.434ms Count:6666}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource:binding Verb:POST Scope:namespace Latency:perc50: 9.179ms, perc90: 100.26ms, perc99: 219.863ms Count:3333}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:pods Subresource: Verb:POST Scope:namespace Latency:perc50: 11.573ms, perc90: 68.031ms, perc99: 205.966ms Count:3333}; threshold: 1s
APIResponsiveness: Top latency metric: {Resource:deployments Subresource: Verb:DELETE Scope:namespace Latency:perc50: 3.56ms, perc90: 18.001ms, perc99: 201.319ms Count:328}; threshold: 1s
```